Changed weight from int to double in Path.

+++ Graph/src/Path.java
@@ -5,7 +5,7 @@ public class Path
 {
 	private Vertex from;
 	private Vertex to;
-	private int path_weight;
+	private double path_weight;
 	boolean isSelected = false;
 	boolean isDirectionEnabled = false;
 	
@@ -33,7 +33,7 @@ public class Path
 		return to;
 	}
 	
-	public int getWeight()
+	public double getWeight()
 	{
 		return path_weight;
 	}	
@@ -65,9 +65,9 @@ public class Path
 		return "Path: from=(" + from.toString() + "), to=(" + to.toString() + "), distance=" + path_weight;
 	}
 	
-	private int weight()
+	private double weight()
 	{
-		return (int)(MapEditor.scale_feet_per_pixel * Math.sqrt(Math.pow((to.getX()-from.getX()), 2) + Math.pow((to.getY()-from.getY()), 2)));
+		return (MapEditor.scale_feet_per_pixel * Math.sqrt(Math.pow((to.getX()-from.getX()), 2) + Math.pow((to.getY()-from.getY()), 2)));
 	}
 	
 	public boolean equals(Object obj)
